CONTROLLER LAYER

+++ health_tracker/src/main/java/org/elsys/health_tracker/service/impl/MealServiceImpl.java
@@ -1,10 +1,13 @@
 package org.elsys.health_tracker.service.impl;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.elsys.health_tracker.controller.resources.MealResource;
 import org.elsys.health_tracker.entity.Meal;
+import org.elsys.health_tracker.exception.DuplicateEntityFieldException;
 import org.elsys.health_tracker.repository.MealRepository;
 import org.elsys.health_tracker.service.MealService;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -34,23 +37,35 @@ public class MealServiceImpl implements MealService {
 
     @Override
     public MealResource create(MealResource mealResource) {
-        Meal  meal = mealRepository.save(MEAL_MAPPER.fromMealResource(mealResource));
-        mealResource.setId(meal.getId());
-        return mealResource;
+        try {
+            Meal  meal = mealRepository.save(MEAL_MAPPER.fromMealResource(mealResource));
+            mealResource.setId(meal.getId());
+            return mealResource;
+        } catch (DataIntegrityViolationException e) {
+            throw new DuplicateEntityFieldException("Meal with name " + mealResource.getName() + " already exists.");
+        }
     }
 
     @Override
     public MealResource update(MealResource mealResource, Long id) {
-        Meal meal = mealRepository.getReferenceById(id);
+        try {
+            Meal meal = mealRepository.getReferenceById(id);
 
-        meal.setName(mealResource.getName());
-        meal.setCalories(mealResource.getCalories());
+            meal.setName(mealResource.getName());
+            meal.setCalories(mealResource.getCalories());
 
-        return MEAL_MAPPER.toMealResource(mealRepository.save(meal));
+            return MEAL_MAPPER.toMealResource(mealRepository.save(meal));
+        } catch (EntityNotFoundException e) {
+            throw new EntityNotFoundException("Unable to find meal with id " + id + ".");
+        }
     }
 
     @Override
     public void delete(Long id) {
-        mealRepository.deleteById(id);
+        if (mealRepository.existsById(id)) {
+            mealRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Unable to find meal with id " + id + ".");
+        }
     }
 }
