+++ health_tracker/src/main/java/org/elsys/health_tracker/service/impl/MealStatServiceImpl.java
@@ -1,11 +1,14 @@
 package org.elsys.health_tracker.service.impl;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.elsys.health_tracker.controller.resources.MealStatResource;
+import org.elsys.health_tracker.controller.resources.UserMealStatResource;
 import org.elsys.health_tracker.entity.MealStat;
+import org.elsys.health_tracker.mapper.DateMapper;
 import org.elsys.health_tracker.repository.MealStatRepository;
-import org.elsys.health_tracker.service.MealService;
 import org.elsys.health_tracker.service.MealStatService;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,16 +20,10 @@ import static org.elsys.health_tracker.mapper.MealStatMapper.MEAL_STAT_MAPPER;
 @RequiredArgsConstructor
 public class MealStatServiceImpl implements MealStatService {
     private final MealStatRepository mealStatRepository;
-    private final MealService mealService;
 
     @Override
     public List<MealStatResource> getAll() {
         return MEAL_STAT_MAPPER.toMealStatResources(mealStatRepository.findAll());
-    }
-
-    @Override
-    public List<MealStatResource> getAllByUserId(Long userId) {
-        return MEAL_STAT_MAPPER.toMealStatResources(mealStatRepository.findAllByUserId(userId));
     }
 
     @Override
@@ -35,24 +32,41 @@ public class MealStatServiceImpl implements MealStatService {
     }
 
     @Override
+    public Optional<List<UserMealStatResource>> getAllByUserId(Long userId) {
+        return mealStatRepository.findAllByUserId(userId).map(MEAL_STAT_MAPPER::toUserMealStatResources);
+    }
+
+    @Override
     public MealStatResource create(MealStatResource mealStatResource) {
-        MealStat mealStat = mealStatRepository.save(MEAL_STAT_MAPPER.fromMealStatResource(mealStatResource));
-        mealStatResource.setId(mealStat.getId());
-        return mealStatResource;
+        try {
+            MealStat mealStat = mealStatRepository.save(MEAL_STAT_MAPPER.fromMealStatResource(mealStatResource));
+            mealStatResource.setId(mealStat.getId());
+            return mealStatResource;
+        } catch (DataIntegrityViolationException e) {
+            throw new EntityNotFoundException("Foreign key constraint violation.");
+        }
     }
 
     @Override
     public MealStatResource update(MealStatResource mealStatResource, Long id) {
-        MealStat mealStat = mealStatRepository.getReferenceById(id);
+        try {
+            MealStat mealStat = mealStatRepository.getReferenceById(id);
 
-        // future consideration needed
-        mealService.update(mealStatResource.getMeal(), mealStat.getMeal().getId());
+            // TODO: add update logic for meal
+            mealStat.setDate(DateMapper.toSQLDate(mealStatResource.getDate()));
 
-        return MEAL_STAT_MAPPER.toMealStatResource(mealStatRepository.save(mealStat));
+            return MEAL_STAT_MAPPER.toMealStatResource(mealStatRepository.save(mealStat));
+        } catch (EntityNotFoundException e) {
+            throw new EntityNotFoundException("Unable to find meal stat with id " + id + ".");
+        }
     }
 
     @Override
     public void delete(Long id) {
-        mealStatRepository.deleteById(id);
+        if (mealStatRepository.existsById(id)) {
+            mealStatRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Unable to find meal stat with id " + id + ".");
+        }
     }
 }
