+++ health_tracker/src/main/java/org/elsys/health_tracker/service/impl/UserServiceImpl.java
@@ -1,10 +1,13 @@
 package org.elsys.health_tracker.service.impl;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.elsys.health_tracker.controller.resources.UserResource;
 import org.elsys.health_tracker.entity.User;
+import org.elsys.health_tracker.exception.DuplicateEntityFieldException;
 import org.elsys.health_tracker.repository.UserRepository;
 import org.elsys.health_tracker.service.*;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,27 +35,29 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserResource create(UserResource userResource) {
-        User user = userRepository.save(USER_MAPPER.fromUserResource(userResource));
-        userResource.setId(user.getId());
-        return userResource;
-    }
-
-    @Override
     public UserResource update(UserResource userResource, Long id) {
-        User user = userRepository.getReferenceById(id);
+       try {
+           User user = userRepository.getReferenceById(id);
 
-        // TODO: I know
-        user.setUsername(userResource.getUsername());
-        user.setPassword(userResource.getPassword());
-        profileService.update(userResource.getProfile(), user.getProfile().getId());
+           // TODO: I know
+           user.setUsername(userResource.getUsername());
+           user.setPassword(userResource.getPassword());
 
-        return USER_MAPPER.toUserResource(userRepository.save(user));
+           return USER_MAPPER.toUserResource(userRepository.save(user));
+       } catch (EntityNotFoundException e) {
+           throw new EntityNotFoundException("Unable to find user with id " + id + ".");
+       } catch (DataIntegrityViolationException e) {
+           throw new DuplicateEntityFieldException("Username already exists.");
+       }
     }
 
 
     @Override
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Unable to find user with id " + id + ".");
+        }
     }
 }
